blockchain/requirements.txt
asn1crypto==0.24.0
blessings==1.6.1
bpython==0.17
certifi==2017.11.5
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
curtsies==0.2.11
Django==2.0.1
djangorestframework==3.7.7
Faker==0.8.9
greenlet==0.4.12
idna==2.6
pycparser==2.18
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
six==1.11.0
text-unidecode==1.1
ua-parser==0.3.6
urllib3==1.22
wcwidth==0.1.7
